rename scraper files and add MMOChamptionScrapper
