+++ src/core/web_scrapers/base.py
@@ -12,6 +12,10 @@ class Scraper(ABC):
     """
 
     def __init__(self, url: str, parser_type: Optional[str] = "html.parser") -> None:
+        """
+        :param url: A complete url to scrape (incluing query parameters).
+        :param parser_type: Which bs4 parser to use. Default = html.parser
+        """
         self.url = url
         self.parser_type = parser_type
         self.page = None
